+++ lib/router.js
@@ -1177,7 +1177,7 @@ router.get('/cyzone/label/:name', lazyloadRouteHandler('./routes/cyzone/label'))
 router.get('/gov/province/:name/:category', lazyloadRouteHandler('./routes/gov/province'));
 router.get('/gov/city/:name/:category', lazyloadRouteHandler('./routes/gov/city'));
 router.get('/gov/statecouncil/briefing', lazyloadRouteHandler('./routes/gov/statecouncil/briefing'));
-router.get('/gov/news/:uid', lazyloadRouteHandler('./routes/gov/news'));
+// router.get('/gov/news/:uid', lazyloadRouteHandler('./routes/gov/news'));
 router.get('/gov/shuju/:caty/:item', lazyloadRouteHandler('./routes/gov/shuju'));
 router.get('/gov/xinwen/tujie/:caty', lazyloadRouteHandler('./routes/gov/xinwen/tujie'));
