+++ lib/v2/gov/maintainer.js
@@ -28,6 +28,7 @@ module.exports = {
     '/moe/s78/:column': ['TonyRL'],
     '/mofcom/article/:suffix+': ['LogicJake'],
     '/moj/aac/news/:type?': ['TonyRL'],
+    '/news/:uid': ['EsuRt'],
     '/nifdc/:path+': ['nczitzk'],
     '/nmpa/:path+': ['TonyRL'],
     '/nopss/:path+': ['nczitzk'],
