+++ lib/v2/gov/radar.js
@@ -892,6 +892,17 @@ module.exports = {
             },
         ],
     },
+    'sousuo.gov.cn': {
+        _name: '中国政府网',
+        kjt: [
+            {
+                title: '政府新闻 - 政策文件',
+                docs: 'https://docs.rsshub.app/government.html#shan-xi-sheng-sheng-ren-min-zheng-fu-sheng-ke-xue-ji-shu-ting',
+                source: ['/s.htm', '/'],
+                target: '/gov/news/zhengce',
+            },
+        ],
+    },
     'stats.gov.cn': {
         _name: '国家统计局',
         www: [
@@ -972,6 +983,30 @@ module.exports = {
     'www.gov.cn': {
         _name: '中国政府网',
         '.': [
+            {
+                title: '政府新闻 - 政务部门',
+                docs: 'https://docs.rsshub.app/government.html#zhong-guo-zheng-fu-wang',
+                source: ['/lianbo/bumen/index.htm', '/'],
+                target: '/gov/news/bm',
+            },
+            {
+                title: '政府新闻 - 滚动新闻',
+                docs: 'https://docs.rsshub.app/government.html#zhong-guo-zheng-fu-wang',
+                source: ['/xinwen/gundong.htm', '/'],
+                target: '/gov/news/gd',
+            },
+            {
+                title: '政府新闻 - 新闻要闻',
+                docs: 'https://docs.rsshub.app/government.html#zhong-guo-zheng-fu-wang',
+                source: ['/yaowen/index.htm', '/'],
+                target: '/gov/news/yw',
+            },
+            {
+                title: '政府新闻 - 国务院新闻',
+                docs: 'https://docs.rsshub.app/government.html#zhong-guo-zheng-fu-wang',
+                source: ['/pushinfo/v150203', '/'],
+                target: '/gov/news/gwy',
+            },
             {
                 title: '最新政策',
                 docs: 'https://docs.rsshub.app/government.html#zhong-guo-zheng-fu-wang',
