+++ lib/v2/gov/router.js
@@ -20,6 +20,7 @@ module.exports = function (router) {
     router.get('/moe/s78/:column', require('./moe/s78'));
     router.get('/mofcom/article/:suffix+', require('./mofcom/article'));
     router.get('/moj/aac/news/:type?', require('./moj/aac/news'));
+    router.get('/news/:uid', require('./news'));
     router.get(/nifdc\/([\w/-]+)?/, require('./nifdc'));
     router.get(/\/nmpa\/([\w][\w/]+)?/, require('./nmpa/generic'));
     router.get(/nopss(\/[\w/-]+)?/, require('./nopss'));
